move error snack bar to top app level so that errors are always shown on every page
