+++ client/src/App.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { MuiThemeProvider } from "@material-ui/core";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 
@@ -6,27 +6,88 @@ import { theme } from "./themes/theme";
 import LandingPage from "./pages/Landing";
 
 import "./App.css";
-import AuthProvider from "./utils/AuthProvider";
+import AuthProvider, { useAuth } from "./utils/AuthProvider";
 import AuthPage from "./pages/AuthPage";
+import NavBar from "./components/NavBar";
+import Profile from "./pages/Profile";
+
+import { makeStyles } from "@material-ui/core/styles";
+import { Snackbar } from "@material-ui/core";
+import { MySnackbarContentWrapper } from "./components/Snackbar";
+
+const useStyles = makeStyles(theme => ({
+    margin: {
+        margin: theme.spacing(1)
+    }
+}));
 
 function App() {
+    const { user } = useAuth();
+    const { authError, setAuthError } = useAuth();
+    const [open, setOpen] = useState(false);
+    const classes = useStyles();
+
+    const handleOpen = () => {
+        setOpen(true);
+    };
+
+    const handleClose = () => {
+        setAuthError(null); //clear auth error
+        setOpen(false);
+    };
+
+    useEffect(() => {
+        if (authError && !open) {
+            handleOpen();
+        }
+    });
+
     return (
         <MuiThemeProvider theme={theme}>
-            <AuthProvider>
-                <BrowserRouter>
-                    <Switch>
-                        <Route exact path="/" component={LandingPage} />
-                        <Route exact path="/login">
-                            <AuthPage name="login" displayName="Log In" />
-                        </Route>
-                        <Route exact path="/signup">
-                            <AuthPage name="signup" displayName="Sign Up" />
-                        </Route>
-                    </Switch>
-                </BrowserRouter>
-            </AuthProvider>
+            <BrowserRouter>
+                <NavBar />
+                <Switch>
+                    {user ? (
+                        <Route exact path="/mypets" component={Profile} />
+                    ) : (
+                        ""
+                    )}
+                    <Route exact path="/" component={LandingPage} />
+                    <Route exact path="/login">
+                        <AuthPage name="login" displayName="Log In" />
+                    </Route>
+                    <Route exact path="/signup">
+                        <AuthPage name="signup" displayName="Sign Up" />
+                    </Route>
+                </Switch>
+                <Snackbar
+                    anchorOrigin={{ vertical: "bottom", horizontal: "right" }}
+                    open={open}
+                    onClose={handleClose}
+                    autoHideDuration={5000}
+                >
+                    <MySnackbarContentWrapper
+                        variant="error"
+                        className={classes.margin}
+                        onClose={handleClose}
+                        message={
+                            (authError &&
+                                authError.data &&
+                                authError.data.errorMsg) ||
+                            ""
+                        }
+                    />
+                </Snackbar>
+            </BrowserRouter>
         </MuiThemeProvider>
     );
 }
 
-export default App;
+function AuthWrappedApp() {
+    return (
+        <AuthProvider>
+            <App />
+        </AuthProvider>
+    );
+}
+export default AuthWrappedApp;
